metiq: updated E2E test
For some reason the fiducials are not recognized at lower resolution currently.
This increases the time for running the tests.
